requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pandas==2.1.3
pydantic==2.5.0
python-dotenv==1.0.0
langchain==0.0.350
langchain-openai==0.0.2
langchain-community==0.0.10
openai==1.3.0
python-multipart==0.0.6
